Fixed the argument type of limit_virtual_memory(): should be unsigned long.

+++ demos/ppl_pips/ppl_pips.cc
@@ -618,7 +618,7 @@ warning(const char* format, ...) {
 #if PPL_HAVE_DECL_RLIMIT_AS
 
 void
-limit_virtual_memory(const unsigned bytes) {
+limit_virtual_memory(const unsigned long bytes) {
   struct rlimit t;
 
   if (getrlimit(RLIMIT_AS, &t) != 0)
@@ -634,7 +634,7 @@ limit_virtual_memory(const unsigned bytes) {
 #else
 
 void
-limit_virtual_memory(unsigned) {
+limit_virtual_memory(unsigned long) {
 }
 
 #endif // !PPL_HAVE_DECL_RLIMIT_AS
